# openapi.yaml
openapi: 3.0.0
info:
  version: '1.1'
  title: Ruokavälitys API
  description: API for both, managing the RV system and it's inventory, and browsing and purchasing products.
...
servers:
  - url: https://rv-backend-dev.herokuapp.com/
    description: Staging environment
  - url: https://rv-backend.herokuapp.com/
    description: Production environment
components:
  securitySchemes:
    normalAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the \"end user facing\" portition of the API.
        Can be obtained with any valid credentials from route `/api/v1/authenticate`.
    adminAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        A JWT access token used for authenticating with the administrations portition of the API.
        Can be obtained with administrator credentials from route `/api/v1/admin/authenticate`.
  schemas:
    Role:
      type: string
      enum:
        - ADMIN
        - USER1
      example: USER
...
    DepositEventBase:
      type: object
      required: 
        - depositId
        - time
        - amount
        - balanceAfter
      properties: 
        depositId: 
          type: integer
        time: 
          type: string
          format: date-time
          example: '2018-12-24T00:00:00.000Z'
        amount: 
          type: number
          example: 20.00
        balanceAfter:
          type: number
          example: 25.00
    DepositEventUser:
      type: object
      required:
        - user
      properties:
        user:
          '$ref': '#/components/schemas/User'
    User: 
      type: object
      required: 
        - userId
        - username
        - fullName
        - email
        - moneyBalance
        - role
      properties: 
        userId: 
          type: integer
          example: 415
        username: 
          type: string
          example: fuxi2001
        fullName: 
          type: string
          example: Fuxi C. S. McStudent
        email: 
          type: string
          example: fuxi.student@helsinki.fi
        moneyBalance: 
          type: number
          example: 20.00
        role:
          '$ref': '#/components/schemas/Role'
...
  responses:
...
    NotAuthorizedErrorResponse:
      description: Your user lacks the required privilege level to perform this action.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_authorized
              message:
                type: string
                example: Not authorized
    NotFoundErrorResponse:
      description: The resource your are requesting does not exist.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - not_found
              message:
                type: string
                example: No product with barcode \'1312767817\' found
    BadRequestErrorResponse:
      description: The request you sent contains invalid values or is missing fields.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - bad_request
                  - invalid_reference
              message:
                type: string
                example: Missing or invalid fields in request
              errors:
                type: array
                items:
                  type: string
                  example: 'boxBarcode should be a numeric 1-14 digit barcode' 
    IdentifierTakenErrorResponse:
      description: Creating a new resource failed, because a resource with the given identifier already exists.
      content:
        application/json:
          schema:
            type: object
            required:
              - error_code
              - message
            properties:
              error_code:
                type: string
                enum:
                  - identifier_taken
              message:
                type: string
                example: Barcode already in use.
paths: 
...
  /api/v1/user:
    get: 
      summary: Get details of the authenticated user
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user:
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
    patch: 
      summary: Update details of the authenticated user
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: User information updated successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user: 
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        409:
          '$ref': '#/components/responses/IdentifierTakenErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              properties: 
                username: 
                  type: string
                  example: 'jäärä2001'
                fullName: 
                  type: string
                  example: 'Jäärä F. S. McConsultant'
                email: 
                  type: string
                  example: 'jaara.fs.constultant@coolcorp.com'
  /api/v1/user/deposit: 
    post: 
      summary: Deposit money to the authenticated user's account
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: Balance added to the account successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - accountBalance
                  - deposit
                properties: 
                  accountBalance: 
                    type: number
                  deposit: 
                    allOf:
                      - '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
      requestBody: 
        required: true
        content: 
          application/json: 
            schema: 
              type: object
              required: 
                - amount
              properties: 
                amount: 
                  type: string
...
  /api/v1/user/depositHistory: 
    get: 
      summary: Get list of past deposit events of the authenticated user
      security:
        - normalAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposits
                properties: 
                  deposits: 
                    type: array
                    items:
                      '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/user/depositHistory/{depositId}: 
    get: 
      summary: Get details of a specific deposit event created by the authenticated user
      security:
        - normalAccessToken: []
      parameters: 
        - name: depositId
          required: true
          in: path
          schema: 
            type: string
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposit
                properties: 
                  deposit: 
                    '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
  /api/v1/user/changePassword:
    post:
      summary: Change the account's password
      security:
        - normalAccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  example: hunter2
      responses:
        204:
          description: Password changed successfully
        400:
          '$ref': '#/components/responses/BadRequestErrorResponse'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
...
  /api/v1/admin/users: 
    get: 
      summary: Get a list of all users
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - users
                properties: 
                  users: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/admin/users/{userId}: 
    get: 
      summary: Get the details of a user
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          required: true
          in: path
          schema: 
            type: integer
            example: 324
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - user
                properties: 
                  user: 
                    '$ref': '#/components/schemas/User'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/users/{userId}/depositHistory: 
    get: 
      summary: Get deposit history of a user
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          required: true
          in: path
          schema: 
            type: integer
            example: 324
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposits
                properties: 
                  deposits: 
                    type: array
                    items: 
                      '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
  /api/v1/admin/users/{userId}/depositHistory/{depositId}: 
    get: 
      summary: Get details of a specific deposit event
      security:
        - adminAccessToken: []
      parameters: 
        - name: userId
          required: true
          in: path
          schema: 
            type: integer
            example: 324
        - name: depositId
          required: true
          in: path
          schema: 
            type: integer
            example: 3015
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposit
                properties: 
                  deposit: 
                    '$ref': '#/components/schemas/DepositEventBase'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...
  /api/v1/admin/depositHistory: 
    get: 
      summary: Get the combined deposit history of all users
      security:
        - adminAccessToken: []
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposits
                properties: 
                  deposits: 
                    type: array
                    items: 
                      allOf:
                        - '$ref': '#/components/schemas/DepositEventBase'
                        - '$ref': '#/components/schemas/DepositEventUser'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
  /api/v1/admin/depositHistory/{depositId}: 
    get: 
      summary: Get details of a specific deposit event
      security:
        - adminAccessToken: []
      parameters: 
        - name: depositId
          required: true
          in: path
          schema: 
            type: integer
            example: 3015
      responses: 
        200: 
          description: Request fulfilled successfully
          content: 
            application/json: 
              schema: 
                type: object
                required: 
                  - deposit
                properties: 
                  deposit: 
                    allOf:
                        - '$ref': '#/components/schemas/DepositEventBase'
                        - '$ref': '#/components/schemas/DepositEventUser'
        401:
          '$ref': '#/components/responses/NotAuthorizedErrorResponse'
        404:
          '$ref': '#/components/responses/NotFoundErrorResponse'
...